requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Data Generation
faker>=18.0.0

# Database
duckdb>=0.8.0

# Machine Learning
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0
category-encoders>=2.6.0
joblib>=1.3.0

# Feature Engineering
feature-engine>=1.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0

# Testing
pytest>=7.0.0
pytest-mock>=3.0.0

# Documentation
jupyter>=1.0.0
jupyterlab>=4.0.0

# API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
redis>=5.0.0
celery>=5.3.0

# Dashboard Dependencies
streamlit>=1.28.0
plotly>=5.17.0
altair>=5.2.0
pydeck>=0.8.0
streamlit-aggrid>=0.3.0
streamlit-authenticator>=0.2.0

# Development Dependencies
httpx>=0.25.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.5.0

# Data & ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Database
sqlalchemy>=2.0.0
duckdb>=0.8.0

# Utils
python-dotenv>=1.0.0
loguru>=0.7.0